Initial commit for the new Annotated Aggregate API
Instead of extending classes or implementing interfaces, aggregates now only
need certain anntoations to demarcate relevant properties.
Tests still need fixing and Sagas will need to undergo a similar makeover.

+++ integrationtests/src/test/java/org/axonframework/integrationtests/jpa/AbstractAnnotatedAggregateRoot_PersistenceTest.java
@@ -53,7 +53,7 @@ public class AbstractAnnotatedAggregateRoot_PersistenceTest {
         SimpleJpaEventSourcedAggregate reloaded = entityManager.find(SimpleJpaEventSourcedAggregate.class, id);
         assertEquals(id, reloaded.getIdentifier());
         assertEquals(2, reloaded.getInvocationCount());
-        assertEquals((Long) 1L, reloaded.getVersion());
+        assertEquals(1L, reloaded.getVersion());
 
         reloaded.doSomething();
